requirements.txt
multiversx-sdk
selenium
python-dotenv
google-generativeai
webdriver-manager
requests
tweepy==4.12.1